+++ java/vector/src/test/java/org/apache/arrow/vector/DirtyRootAllocator.java
@@ -17,9 +17,10 @@
 
 package org.apache.arrow.vector;
 
-import org.apache.arrow.memory.ArrowBuf;
 import org.apache.arrow.memory.BufferManager;
 import org.apache.arrow.memory.RootAllocator;
+
+import siren.io.netty.buffer.ArrowBuf;
 
 /**
  * Root allocator that returns buffers pre-filled with a given value.<br>
@@ -35,12 +36,12 @@ public class DirtyRootAllocator extends RootAllocator {
   }
 
   @Override
-  public ArrowBuf buffer(long size) {
+  public ArrowBuf buffer(int size) {
     return buffer(size, null);
   }
 
   @Override
-  public ArrowBuf buffer(long size, BufferManager manager) {
+  public ArrowBuf buffer(int size, BufferManager manager) {
     ArrowBuf buffer = super.buffer(size, manager);
     // contaminate the buffer
     for (int i = 0; i < buffer.capacity(); i++) {
